[webkitcorepy] Attempt to terminate stuck processes before killing them https://bugs.webkit.org/show_bug.cgi?id=217632 <rdar://problem/70222803>
Reviewed by Dewei Zhu.

* Scripts/libraries/webkitcorepy/webkitcorepy/__init__.py: Bump version.
* Scripts/libraries/webkitcorepy/webkitcorepy/subprocess_utils.py:
(run): Killed processes will sometimes not have stdout or stderr, while terminated processes often will.


Canonical link: https://commits.webkit.org/230382@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268373 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/libraries/webkitcorepy/webkitcorepy/subprocess_utils.py
@@ -118,12 +118,21 @@ else:
                 stdout, stderr = process.communicate(input)
 
         except Timeout.Exception:
-            process.kill()
-            stdout, stderr = process.communicate()
+            process.terminate()
+            deadline = time.time() + .5
+            while process.poll() is None and time.time() < deadline:
+                time.sleep(.1)
+            if process.poll() is None:
+                process.kill()
+                stdout, stderr = '', ''
+            else:
+                stdout, stderr = process.communicate()
             raise TimeoutExpired(popenargs[0], timeout, output=decode(stdout), stderr=decode(stderr))
+
         except BaseException:
             process.kill()
             raise
+
         finally:
             process.wait()
